refactor: better fs structure
