+++ src/bindings.ts
@@ -1,9 +1,12 @@
 import { View } from 'react-native';
+import type { KeyboardControllerModule } from './types';
 
 const NOOP = () => {};
-export const KeyboardController = {
+export const KeyboardController: KeyboardControllerModule = {
   setDefaultMode: NOOP,
   setInputMode: NOOP,
+  addListener: NOOP,
+  removeListeners: NOOP,
 };
 export const KeyboardEvents = {
   addListener: () => NOOP,
