+++ src/index.ts
@@ -1,6 +1,6 @@
 import './monkey-patch';
 
-export * from './native';
+export * from './bindings';
 export * from './animated';
 export * from './replicas';
 export * from './context';
